Fixed the GPS temperature

+++ src/TADC.cpp
@@ -289,6 +289,7 @@ int TADC::SetValuesFromPointers(unsigned short *pevent, string file_format)
 		gps_long.push_back(*(unsigned long long*)&evdu[file_shift + EVT_LONGITUDE]);
 		gps_lat.push_back(*(unsigned long long*)&evdu[file_shift + EVT_LATITUDE]);
 		gps_alt.push_back(*(unsigned long long*)&evdu[file_shift + EVT_ALTITUDE]);
+		// ToDo: I think this casting is a bug, and it should be unsigned int
 		gps_temp.push_back(*(unsigned long long*)&evdu[file_shift + EVT_GPS_TEMP]);
 
 //		digi_ctrl.push_back(vector<unsigned short>());
@@ -486,7 +487,7 @@ int TADC::SetValuesFromPointers_fv2(unsigned short *pevent, string file_format)
 		gps_lat.push_back(*(unsigned long long*)&evdu[EVT_LATITUDE-1]);
 		gps_alt.push_back(*(unsigned long long*)&evdu[EVT_ALTITUDE-1]);
 		// ToDo: Now this is float, so just 2 bytes. How to fix it?
-		gps_temp.push_back(*(unsigned long long*)&evdu[EVT_TEMPERATURE]);
+		gps_temp.push_back(*(unsigned int*)&evdu[EVT_TEMPERATURE]);
 
 		// ToDo: Add seconds since sunday, week, utc offset, modes and work on alarms
